+++ constant/constant.go
@@ -2,7 +2,7 @@ package constant
 
 const (
 	Mangal    = "mangal"
-	Version   = "3.9.1"
+	Version   = "3.10.0"
 	UserAgent = "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/104.0.0.0 Safari/537.36"
 )
 
@@ -10,15 +10,6 @@ const (
 	TempPrefix  = Mangal + "Temp_"
 	CachePrefix = Mangal + "Cache_"
 )
-
-const AssciiArtLogo = `
-                                _   _____ 
-  /\/\   __ _ _ __   __ _  __ _| | |___ / 
- /    \ / _' | '_ \ / _' |/ _' | |   |_ \ 
-/ /\/\ \ (_| | | | | (_| | (_| | |  ___) |
-\/    \/\__,_|_| |_|\__, |\__,_|_| |____/
-                    |___/
-`
 
 const (
 	SearchMangaFn   = "SearchManga"