+++ tui/keymap.go
@@ -11,6 +11,7 @@ type statefulKeymap struct {
 
 	quit, forceQuit,
 	selectOne, selectAll, selectVolume, clearSelection,
+	anilistSelect,
 	remove,
 	redownloadFailed,
 	confirm,
@@ -77,6 +78,10 @@ func newStatefulKeymap() *statefulKeymap {
 		redownloadFailed: k(
 			keys("r"),
 			help("r", "redownload failed"),
+		),
+		anilistSelect: k(
+			keys("a"),
+			help("a", "select anilist manga"),
 		),
 		openFolder: k(
 			keys("o"),
@@ -150,7 +155,9 @@ func (k *statefulKeymap) help() ([]key.Binding, []key.Binding) {
 		return to2(h(k.confirm, k.back, k.openURL))
 	case chaptersState:
 		download := withDescription(k.confirm, "download selected")
-		return h(k.read, k.selectOne, k.selectAll, download, k.back), h(k.read, k.selectOne, k.selectAll, k.clearSelection, k.openURL, download, k.selectVolume, k.back)
+		return h(k.read, k.selectOne, k.selectAll, download, k.back), h(k.read, k.selectOne, k.selectAll, k.clearSelection, k.openURL, download, k.selectVolume, k.anilistSelect, k.back)
+	case anilistSelectState:
+		return to2(h(k.confirm, k.openURL, k.back))
 	case confirmState:
 		return to2(h(k.confirm, k.back, k.quit))
 	case readState: