+++ tui/state.go
@@ -11,6 +11,7 @@ const (
 	searchState
 	mangasState
 	chaptersState
+	anilistSelectState
 	confirmState
 	readState
 	downloadState